added new files, interfaces, adapting with the changes the team discuss on morning 6th June Friday
