+++ src/contexts/LearningContext/application/useCases/GetAllCoursesUseCase.ts
@@ -1,7 +1,8 @@
 import { ICourseRepository } from "../../domain/interfaces/ICourseRepository";
 import { Course } from "../../domain/aggregates/Course";
+import { IUseCase } from "../../domain/interfaces/IUseCase";
 
-export class GetAllCoursesUseCase {
+export class GetAllCoursesUseCase implements IUseCase<void, Course[]> {
   constructor(private repo: ICourseRepository) {}
   async execute(): Promise<Course[]> {
     return await this.repo.findAll();
