+++ src/contexts/LearningContext/domain/interfaces/ICourseService.ts
@@ -4,4 +4,5 @@ import { CourseDTO } from "../dtos/CourseDTO";
 
 export interface ICourseService {
   getAllCourses(): Promise<CourseDTO[]>;
+  create(courseDto: CourseDTO): Promise<CourseDTO>;
 }
