+++ src/contexts/LearningContext/domain/interfaces/IUseCase.ts
@@ -1,0 +1,3 @@
+export interface IUseCase<Params, Response> {
+  execute(params?: Params): Promise<Response> | Response;
+}
