+++ src/contexts/LearningContext/presentation/http/controllers/CourseController.ts
@@ -1,11 +1,11 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { Request, Response } from "express";
-import { CourseService } from "../../../infraestructure/services/CourseService";
 import { ICourseController } from "@/contexts/LearningContext/domain/interfaces/ICourseController";
 import { CourseDTO } from "@/contexts/LearningContext/domain/dtos/CourseDTO";
+import { ICourseService } from "@/contexts/LearningContext/domain/interfaces/ICourseService";
 
 export class CourseController implements ICourseController {
-  constructor(private readonly courseService: CourseService) {}
+  constructor(private readonly courseService: ICourseService) {}
 
   public getAllCourses = async (req: Request, res: Response): Promise<void> => {
     try {
@@ -22,8 +22,9 @@ export class CourseController implements ICourseController {
       const dto = req.body as CourseDTO;
       const result = await this.courseService.create(dto);
       return res.status(201).json(result);
-    } catch (error: any) {
-      return res.status(400).json({ message: error.message });
+    } catch (error) {
+      console.error("Error in CourseController.create:", error);
+      return res.status(500).json({ message: "Internal server error" });
     }
   };
 }
