+++ src/contexts/LearningContext/presentation/http/routes/CourseRoutes.ts
@@ -1,18 +1,5 @@
 import { Router } from "express";
-import { CourseController } from "../controllers/CourseController";
-import { CourseRepository } from "../../../infraestructure/database/CourseRepository";
-import { CourseService } from "../../../infraestructure/services/CourseService";
-import { GetAllCoursesUseCase } from "../../../application/useCases/GetAllCoursesUseCase";
-import { CreateCourseUseCase } from "@/contexts/LearningContext/application/useCases/CreateCourseUseCase";
-
-const courseRepository = new CourseRepository();
-const getAllCoursesUseCase = new GetAllCoursesUseCase(courseRepository);
-const createCourseUseCase = new CreateCourseUseCase(courseRepository);
-const courseService = new CourseService(
-  getAllCoursesUseCase,
-  createCourseUseCase
-);
-const controller = new CourseController(courseService);
+import { controller } from "../controllers/CourseMain";
 
 const courseRouter = Router();
